requirements.txt
matplotlib==3.10.8
numpy==2.4.2
Pillow==12.1.1
streamlit==1.54.0
streamlit_drawable_canvas==0.9.3
tensorflow==2.20.0
